rename "server" -> "manifoldserver"
