+++ cmd/server.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/hanoi-radix/manifold-towers-server/internal/server"
+import "github.com/hanoi-radix/manifold-towers-server/internal/manifoldserver"
 
 func main() {
 	manifoldserver.Start()
